+++ sdk/resourcemanager/recoveryservices/armrecoveryservicesbackup/backupjobs_client.go
@@ -45,7 +45,7 @@ func NewBackupJobsClient(subscriptionID string, credential azcore.TokenCredentia
 
 // NewListPager - Provides a pageable list of jobs.
 //
-// Generated from API version 2024-04-01
+// Generated from API version 2024-07-30-preview
 //   - vaultName - The name of the recovery services vault.
 //   - resourceGroupName - The name of the resource group where the recovery services vault is present.
 //   - options - BackupJobsClientListOptions contains the optional parameters for the BackupJobsClient.NewListPager method.
@@ -98,7 +98,7 @@ func (client *BackupJobsClient) listCreateRequest(ctx context.Context, vaultName
 	if options != nil && options.SkipToken != nil {
 		reqQP.Set("$skipToken", *options.SkipToken)
 	}
-	reqQP.Set("api-version", "2024-04-01")
+	reqQP.Set("api-version", "2024-07-30-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil