+++ sdk/resourcemanager/recoveryservices/armrecoveryservicesbackup/backuppolicies_client.go
@@ -46,7 +46,7 @@ func NewBackupPoliciesClient(subscriptionID string, credential azcore.TokenCrede
 // NewListPager - Lists of backup policies associated with Recovery Services Vault. API provides pagination parameters to
 // fetch scoped results.
 //
-// Generated from API version 2024-04-01
+// Generated from API version 2024-07-30-preview
 //   - vaultName - The name of the recovery services vault.
 //   - resourceGroupName - The name of the resource group where the recovery services vault is present.
 //   - options - BackupPoliciesClientListOptions contains the optional parameters for the BackupPoliciesClient.NewListPager method.
@@ -96,7 +96,7 @@ func (client *BackupPoliciesClient) listCreateRequest(ctx context.Context, vault
 	if options != nil && options.Filter != nil {
 		reqQP.Set("$filter", *options.Filter)
 	}
-	reqQP.Set("api-version", "2024-04-01")
+	reqQP.Set("api-version", "2024-07-30-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil