+++ sdk/resourcemanager/recoveryservices/armrecoveryservicesbackup/backupprotectableitems_client.go
@@ -46,7 +46,7 @@ func NewBackupProtectableItemsClient(subscriptionID string, credential azcore.To
 // NewListPager - Provides a pageable list of protectable objects within your subscription according to the query filter and
 // the pagination parameters.
 //
-// Generated from API version 2024-04-01
+// Generated from API version 2024-07-30-preview
 //   - vaultName - The name of the recovery services vault.
 //   - resourceGroupName - The name of the resource group where the recovery services vault is present.
 //   - options - BackupProtectableItemsClientListOptions contains the optional parameters for the BackupProtectableItemsClient.NewListPager
@@ -100,7 +100,7 @@ func (client *BackupProtectableItemsClient) listCreateRequest(ctx context.Contex
 	if options != nil && options.SkipToken != nil {
 		reqQP.Set("$skipToken", *options.SkipToken)
 	}
-	reqQP.Set("api-version", "2024-04-01")
+	reqQP.Set("api-version", "2024-07-30-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil