+++ sdk/resourcemanager/recoveryservices/armrecoveryservicesbackup/backupprotectioncontainers_client.go
@@ -45,7 +45,7 @@ func NewBackupProtectionContainersClient(subscriptionID string, credential azcor
 
 // NewListPager - Lists the containers registered to Recovery Services Vault.
 //
-// Generated from API version 2024-04-01
+// Generated from API version 2024-07-30-preview
 //   - vaultName - The name of the recovery services vault.
 //   - resourceGroupName - The name of the resource group where the recovery services vault is present.
 //   - options - BackupProtectionContainersClientListOptions contains the optional parameters for the BackupProtectionContainersClient.NewListPager
@@ -96,7 +96,7 @@ func (client *BackupProtectionContainersClient) listCreateRequest(ctx context.Co
 	if options != nil && options.Filter != nil {
 		reqQP.Set("$filter", *options.Filter)
 	}
-	reqQP.Set("api-version", "2024-04-01")
+	reqQP.Set("api-version", "2024-07-30-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil