+++ sdk/resourcemanager/recoveryservices/armrecoveryservicesbackup/polymorphic_helpers.go
@@ -583,6 +583,8 @@ func unmarshalWorkloadProtectableItemClassification(rawMsg json.RawMessage) (Wor
 		b = &AzureVMWorkloadProtectableItem{}
 	case "HanaHSRContainer":
 		b = &AzureVMWorkloadSAPHanaHSRProtectableItem{}
+	case "HanaScaleoutContainer":
+		b = &AzureVMWorkloadSAPHanaScaleoutProtectableItem{}
 	case "IaaSVMProtectableItem":
 		b = &IaaSVMProtectableItem{}
 	case "Microsoft.ClassicCompute/virtualMachines":