+++ sdk/resourcemanager/recoveryservices/armrecoveryservicesbackup/protectablecontainers_client.go
@@ -45,7 +45,7 @@ func NewProtectableContainersClient(subscriptionID string, credential azcore.Tok
 
 // NewListPager - Lists the containers that can be registered to Recovery Services Vault.
 //
-// Generated from API version 2024-04-01
+// Generated from API version 2024-07-30-preview
 //   - vaultName - The name of the recovery services vault.
 //   - resourceGroupName - The name of the resource group where the recovery services vault is present.
 //   - options - ProtectableContainersClientListOptions contains the optional parameters for the ProtectableContainersClient.NewListPager
@@ -100,7 +100,7 @@ func (client *ProtectableContainersClient) listCreateRequest(ctx context.Context
 	if options != nil && options.Filter != nil {
 		reqQP.Set("$filter", *options.Filter)
 	}
-	reqQP.Set("api-version", "2024-04-01")
+	reqQP.Set("api-version", "2024-07-30-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil